requirements.txt
alive-progress==3.1.4
numpy==1.23.5
optuna==3.2.0
pandas==2.0.0
plotly==5.14.1
python-dotenv==1.0.0
scikit-learn==1.2.2
tekore==5.0.0
xgboost==1.7.6
mysql-connector-python==8.0.33